Refactor CAN signal definitions and update ECU communication
- Adjusted ECU temperature calculation in high_receive function to apply a -40 offset for brake diagnostics.
- Enhanced can_high.yaml with shared constants, ECU definitions, and updated protocol specifications.
- Modified can_signals_schema.py to read metadata from can_high.yaml and merge protocols from all bus-specific YAML files.
- Updated etrike_high.dbc to reflect changes in RT_STATE_RPT and STEER_DIAG message definitions.
- Revised CAN validator to expect updated DLC for RT_STATE_RPT and added new signals for telemetry.
- Corrected steering angle calculations in RtSteeringController and SyntreeEpsc to use the new offset.
- Updated simulation tests to reflect changes in steering angle encoding and CAN message definitions.
- Added compatibility test for Autoware documentation against CAN signal definitions.
- Introduced a .env.example file for backend configuration in the debug tool.

+++ autoware_vehicle_bridge/src/vehicle_bridge_node.cpp
@@ -816,7 +816,7 @@ void VehicleBridgeNode::publish_vehicle_reports(const struct can_frame & frame)
 
     case 0x310: {  // STEER_DIAG — v0.0.4 EPS-C telemetry
       if (frame.len < 8) break;
-      int16_t angle_raw = (int16_t)((uint16_t)frame.data[0] << 8 | frame.data[1]);
+      uint16_t angle_raw = (uint16_t)frame.data[0] << 8 | frame.data[1];
       float steer_deg = (angle_raw - 30000) * 0.1f;  // offset=-3000, 0.1°/bit
       steer_angle_rad_ = steer_deg * M_PI / 180.0f;  // cache for odometry
       last_steer_time_ = now();
